Risk Analytics Case Study
This case study aims to develop a fundamental understanding of risk analytics in banking and financial services. The goal is to explore how data is utilized to minimize the risk of financial losses when lending to customers. The study follows a structured analysis approach, encompassing various steps from data cleaning to correlation matrix exploration.